Refactored Spotify Player into new component directory
