+++ src/components/MusicPlayer/Spotify/index.jsx
@@ -1,0 +1,3 @@
+import SpotifyPlayer from "./SpotifyPlayer";
+
+export { SpotifyPlayer };
